+++ examples/src/examples/shaders/trees.example.mjs
@@ -181,6 +181,7 @@ assetListLoader.load(() => {
     // ------ End of shader chunks ------
 
     // apply all these chunks to the tree material
+    meshInstance.material.shaderChunks.APIVersion = pc.CHUNKAPI_2_8;
     meshInstance.material.shaderChunks.glsl.set('diffusePS', diffusePS);
     meshInstance.material.shaderChunks.glsl.set('litUserMainEndPS', litUserMainEndPS);
     meshInstance.material.shaderChunks.glsl.set('litUserDeclarationPS', litUserDeclarationPS);
@@ -188,6 +189,7 @@ assetListLoader.load(() => {
 
     // create a ground material - all chunks apart from swaying in the wind, so fog and color blending
     const groundMaterial = new pc.StandardMaterial();
+    groundMaterial.shaderChunks.APIVersion = pc.CHUNKAPI_2_8;
     groundMaterial.shaderChunks.glsl.set('diffusePS', diffusePS);
     groundMaterial.shaderChunks.glsl.set('litUserMainEndPS', litUserMainEndPS);
     groundMaterial.shaderChunks.glsl.set('litUserDeclarationPS', litUserDeclarationPS);
