+++ src/scene/shader-lib/programs/shader-generator-shader.js
@@ -1,9 +1,8 @@
 import { hashCode } from '../../../core/hash.js';
 import { SEMANTIC_ATTR15, SEMANTIC_BLENDINDICES, SEMANTIC_BLENDWEIGHT, SHADERLANGUAGE_GLSL, SHADERLANGUAGE_WGSL } from '../../../platform/graphics/constants.js';
 import { ShaderDefinitionUtils } from '../../../platform/graphics/shader-definition-utils.js';
-import { shaderChunksWGSL } from '../chunks-wgsl/chunks-wgsl.js';
-import { shaderChunks } from '../chunks-glsl/chunks.js';
 import { ShaderGenerator } from './shader-generator.js';
+import { ShaderUtils } from '../shader-utils.js';
 
 class ShaderGeneratorShader extends ShaderGenerator {
     generateKey(options) {
@@ -95,10 +94,8 @@ class ShaderGeneratorShader extends ShaderGenerator {
             meshBindGroupFormat: desc.meshBindGroupFormat
         };
 
-        const chunks = wgsl ? shaderChunksWGSL : shaderChunks;
-        const sharedIncludes = new Map(Object.entries({
-            ...chunks  // default chunks
-        }));
+        // includes - default chunks
+        const sharedIncludes = new Map(wgsl ? ShaderUtils.shaderChunks.wgsl : ShaderUtils.shaderChunks.glsl);
 
         this.createAttributesDefinition(definitionOptions, options);
         this.createVertexDefinition(definitionOptions, options, sharedIncludes, wgsl);
