+++ src/scene/shader-lib/shader-utils.js
@@ -5,11 +5,11 @@ import { Debug } from '../../core/debug.js';
 import { ShaderGenerator } from './programs/shader-generator.js';
 import { ShaderPass } from '../shader-pass.js';
 import { SHADERLANGUAGE_GLSL, SHADERLANGUAGE_WGSL } from '../../platform/graphics/constants.js';
+import { ShaderChunks } from './shader-chunks.js';
 
 /**
  * @import { GraphicsDevice } from '../../platform/graphics/graphics-device.js'
  * @import { ShaderProcessorOptions } from '../../platform/graphics/shader-processor-options.js'
- * @import { CameraShaderParams } from '../camera-shader-params.js'
  * @import { Material, ShaderVariantParams } from '../materials/material.js'
  */
 
@@ -30,6 +30,13 @@ class ShaderGeneratorPassThrough extends ShaderGenerator {
 }
 
 class ShaderUtils {
+    /**
+     * Object containing all default shader chunks used by shader generators.
+     *
+     * @type {ShaderChunks}
+     */
+    static shaderChunks = new ShaderChunks();
+
     /**
      * Creates a shader. When the active graphics device is WebGL, the provided GLSL vertex and
      * fragment source code is used. For WebGPU, if WGSL vertex and fragment source code is
